iScroll plugin and Link Rel plugin got mixed up
